Fixed problem with Dark-mode button
