+++ src/components/NavBar/Nav.js
@@ -1,7 +1,7 @@
 import '../../css/Nav.css';
 import Dropdown from './Dropdown';
 import NavButton from './NavButton.js';
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import { GiHamburgerMenu } from 'react-icons/gi'
 
 //import logos
@@ -72,59 +72,6 @@ function Nav({ setState, tabIndex }) {
       ["Fotoalbum", "https://d-sektionen.se/filarkiv/fotoalbum/"]
     ];
 
-
-    const [darkMode, setDarkMode] = useState(() => {
-      const storedDarkMode = localStorage.getItem('isDarkMode');
-      return storedDarkMode ? JSON.parse(storedDarkMode) : false;
-    });
-
-    useEffect(() => {
-      const isDarkMode = JSON.parse(localStorage.getItem('isDarkMode'));
-      setDarkMode(isDarkMode);
-      console.log('Retrieved from localStorage:', isDarkMode);
-    }, []);
-    
-    useEffect(() => {
-      console.log('Setting dark mode:', darkMode);
-      localStorage.setItem('isDarkMode', JSON.stringify(darkMode));
-      const root = document.documentElement;
-      if (darkMode) {
-        root.style.setProperty('--text-color', 'var(--text-color-dark)');
-        root.style.setProperty('--bg-color', 'var(--bg-color-dark)');
-        root.style.setProperty('--header-bg-color', 'var(--header-bg-color-dark)');
-        root.style.setProperty('--navbar-bg-color', 'var(--navbar-bg-color-dark)');
-        root.style.setProperty('--navbar-text-color', 'var(--navbar-text-color-dark)');
-        root.style.setProperty('--footer-text-color', 'var(--footer-text-color-dark)');
-        root.style.setProperty('--footer-bg-color', 'var(--footer-bg-color-dark)');
-        root.style.setProperty('--widget-title-bg-color', 'var(--widget-title-bg-color-dark)');
-        root.style.setProperty('--widget-title-border-color', 'var(--widget-title-border-color-dark)');
-        root.style.setProperty('--display-light-logo', 'none');
-        root.style.setProperty('--display-dark-logo', 'inline-block');
-        root.style.setProperty('--button-bg-color', 'var(--button-bg-color-dark)');
-        root.style.setProperty('--button-text-color', 'var(--button-text-color-dark)');
-        root.style.setProperty('--navbar-dropdown-bg-color', 'var(--navbar-dropdown-bg-color-dark)');
-        root.style.setProperty('--utskott-bg-color', 'var(--utskott-text-color-dark)');
-        root.style.setProperty('--utskott-text-color', 'var(--utskott-bg-color-dark)');
-      } else {
-        root.style.setProperty('--text-color', 'var(--text-color-light)');
-        root.style.setProperty('--bg-color', 'var(--bg-color-light)');
-        root.style.setProperty('--header-bg-color', 'var(--header-bg-color-light)');
-        root.style.setProperty('--navbar-bg-color', 'var(--navbar-bg-color-light)');
-        root.style.setProperty('--navbar-text-color', 'var(--navbar-text-color-light)');
-        root.style.setProperty('--footer-text-color', 'var(--footer-text-color-light)');
-        root.style.setProperty('--footer-bg-color', 'var(--footer-bg-color-light)');
-        root.style.setProperty('--widget-title-bg-color', 'var(--widget-title-bg-color-light)');
-        root.style.setProperty('--widget-title-border-color', 'var(--widget-title-border-color-light)');
-        root.style.setProperty('--display-light-logo', 'inline-block');
-        root.style.setProperty('--display-dark-logo', 'none');
-        root.style.setProperty('--button-bg-color', 'var(--button-bg-color-light)');
-        root.style.setProperty('--button-text-color', 'var(--button-text-color-light)');
-        root.style.setProperty('--navbar-dropdown-bg-color', 'var(--navbar-dropdown-bg-color-light)');
-        root.style.setProperty('--utskott-bg-color', 'var(--utskott-text-color-light)');
-        root.style.setProperty('--utskott-text-color', 'var(--utskott-bg-color-light)');
-      }
-    }, [darkMode]);
-
     return (
       <>
         <div className='hamburger narrow-nav' tabIndex="0" onClick={handleClick}>
@@ -138,9 +85,6 @@ function Nav({ setState, tabIndex }) {
             <NavButton title="Medlemstjänster" link="https://medlem.d-sektionen.se" />
             <NavButton title="Företag" link="/page/foretag" />
             <Dropdown title="Extra" items={dropdownItemsExtra}></Dropdown>
-            <button type="button" id="dmode" onClick={() => setDarkMode(!darkMode)}>
-              {darkMode ? '!(D-mode)' : 'D-mode'}
-            </button>
           </nav>
       </>
     );
