+++ src/css/Header.css
@@ -22,6 +22,13 @@
         display: none;
     }
 
+    #dmode {
+        font-size: 1.1em;
+        right: 0;
+        left: auto;
+        text-align: right;
+    }
+
 }
 
 .logo {
@@ -39,6 +46,13 @@
         border-right: 40px solid black;
         background-color: black;
         /* border-radius: 5vh; */
+    }
+
+    #dmode {
+        font-size: 0.75em;
+        text-align: left;
+        right: auto;
+        left: 0;
     }
 }
 #logoWhite {
@@ -69,4 +83,17 @@
 .logo {
     height: 90%;
     object-fit: contain;
+}
+
+#dmode {
+    border: solid 1px;
+    border-radius: 2px;
+    padding: 10px 20px;
+    text-decoration: none;
+    display: inline-block;
+    margin: 4px 2px;
+    cursor: pointer;
+    background-color: var(--button-bg-color);
+    color: var(--button-text-color);
+    position: absolute;
 }
