+++ src/css/Nav.css
@@ -92,7 +92,7 @@
     .nav-button {
         margin: 0 6px 0 6px;
         padding: 0 14px 0 14px;
-        font-size: 1.12em;
+        font-size: 1.1em;
         height: 32px;
         line-height: 32px;
         background-color: var(--navbar-bg-color);
@@ -377,17 +377,4 @@ button {
     height: 48px;
     background-color: lightgray;
     text-align: left;
-}
-
-#dmode{
-  border: none;
-  padding: 10px 20px;
-  text-align: center;
-  text-decoration: none;
-  display: inline-block;
-  font-size: 16px;
-  margin: 4px 2px;
-  cursor: pointer;
-  background-color: var(--button-bg-color);
-  color: var(--button-text-color);
 }
